// src/types.ts
export enum ProjectStatus {
  Planning = 'Planning',
  InProgress = 'In Progress',
  Completed = 'Completed',
  OnHold = 'On Hold',
}

export enum TaskStatus {
  ToDo = 'To Do',
  InProgress = 'In Progress',
  Done = 'Done',
}

export interface TeamMember {
  id: string;
  name: string;
  email: string;
  role: string;
}

export interface Client {
  id: string;
  name: string;
  contactPerson: string;
  email: string;
  phone: string;
  location: string;
  createdAt: string; // ISO String
}

export interface Task {
  id: string;
  description: string;
  teamMemberId: string;
  dueDate: string;
  status: TaskStatus;
  sharedWithClient?: boolean;
  notes?: string;
}

export interface Project {
  id:string;
  name: string;
  description: string;
  clientId: string;
  teamMemberIds: string[];
  startDate: string;
  endDate: string;
  status: ProjectStatus;
  tasks: Task[];
}